Improved/Removed usage of old c-style code

+++ OpenGLOrbiter/classes/Main.cpp
@@ -37,8 +37,8 @@ int main () {
 #endif
     
 	//TODO: Add recognition of user screen size and adjust the window size accordingly.
-    const int WindowWidth = 1800*(2.f/3.f);
-    const int WindowHeight = 1350*(2.f/3.f);
+    const int WindowWidth = (int) (1800*(2.f/3.f));
+    const int WindowHeight = (int) (1350*(2.f/3.f));
     
     GLFWwindow* window = glfwCreateWindow (WindowWidth, WindowHeight, "OpenGLOrbiterV0-0-1", NULL, NULL);
     if (!window) {
